Use new querists in checker; run it on dev pushes
