+++ .github/workflows/constants.yml
@@ -1,6 +1,9 @@
 name: CI
 
 on:
+  push:
+    branches:
+      - "dev*"
   schedule:
     - cron: '0 0 * * 1'
